requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
schedule>=1.2.0
requests>=2.31.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
